Se agregan cositas

+++ src/main/java/TpFinalProgramacion/Repositories/SuppliersRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface SuppliersRepository extends JpaRepository<Suppliers, Long> {
-    SupplierDTO getSupplierById(Long id);
+public interface SuppliersRepository extends JpaRepository<Suppliers, Integer> {
+    SupplierDTO getSupplierById(int id);
 }
