+++ src/main/java/TpFinalProgramacion/Services/SuppliersService.java
@@ -2,13 +2,14 @@ package TpFinalProgramacion.Services;
 
 import TpFinalProgramacion.DTOs.SupplierDTO;
 import org.springframework.http.ResponseEntity;
-import java.util.List;
+import org.springframework.stereotype.Service;
 
+import java.util.List;
 public interface SuppliersService {
     List<SupplierDTO> findAll();
-    SupplierDTO getSupplierById(Long id);
+    SupplierDTO getSupplierById(int id);
     ResponseEntity<Object> createSupplier(String name, String email, Integer phoneNumber);
-    ResponseEntity<Object> deleteSupplier(Long id);
+    ResponseEntity<Object> deleteSupplier(int id);
 
 }
